+++ IntegrationTests/TestTransactions.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using IntegrationTests.Support;
+using Iov42sdk.Connection;
 using Iov42sdk.Models;
 using Iov42sdk.Models.Transactions;
 using Iov42sdk.Models.Transfers;
@@ -27,30 +28,26 @@ namespace IntegrationTests
         }
 
         [TestMethod]
-        public async Task ShouldFetchTransactions()
+        public async Task ShouldFetchQuantifiableCreditTransactions()
         {
-            var horseId = _test.CreateUniqueId("horse");
-            await _test.Client.CreateUniqueAssetType(horseId);
-            
-            var trevorId = _test.CreateUniqueId("trevor");
-            var _ = await _test.Client.CreateUniqueAsset(trevorId, horseId);
-            
             var gbpId = _test.CreateUniqueId("GBP");
-            await _test.Client.CreateQuantifiableAssetType(gbpId, 2);
+            var writeResult = await _test.Client.CreateQuantifiableAssetType(gbpId, 2);
+            Assert.IsTrue(writeResult.Success);
             
-            var account = _test.CreateUniqueId("AccountGBP");
-            await _test.Client.CreateQuantifiableAccount(account, gbpId, 1000);
+            var account = _test.CreateUniqueId("TestAccountGBP");
+            writeResult = await _test.Client.CreateQuantifiableAccount(account, gbpId, 1000);
+            Assert.IsTrue(writeResult.Success);
             
-            var __ = await _test.Client.AddBalance(account, gbpId, 50);
+            writeResult = await _test.Client.AddBalance(account, gbpId, 50);
+            Assert.IsTrue(writeResult.Success);
             
             var bruceClient = new IntegrationTestCreation();
-            var bruceAccount = bruceClient.CreateUniqueId("AccountGBP");
-            await bruceClient.Client.CreateQuantifiableAccount(bruceAccount, gbpId, 1000);
+            var bruceAccount = bruceClient.CreateUniqueId("BruceAccountGBP");
+            writeResult = await bruceClient.Client.CreateQuantifiableAccount(bruceAccount, gbpId, 1000);
+            Assert.IsTrue(writeResult.Success);
             
-            var uniqueTransfer = _test.Client.CreateOwnershipTransfer(trevorId, horseId, _test.Identity.Id, bruceClient.Identity.Id);
             var quantityTransfer = _test.Client.CreateQuantityTransfer(account, bruceAccount, gbpId, 10);
-
-            var body = new TransfersBody(quantityTransfer, uniqueTransfer);
+            var body = new TransfersBody(quantityTransfer);
             var bodyText = body.Serialize();
             var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
                 new[]
@@ -58,43 +55,34 @@ namespace IntegrationTests
                     _test.Client.GenerateAuthorisation(bodyText),
                     bruceClient.Client.GenerateAuthorisation(bodyText)
                 });
-            var ____ = await _test.Client.Write(transferRequest);
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
             
-            // Currently only supports credits
-            // var response = await _test.Client.GetTransactions(horseId, trevorId);
-            var response = await _test.Client.GetTransactions(gbpId, account);
-            
-            Assert.IsNotNull(response);
-            Assert.IsTrue(response.Success);
-            Assert.IsNull(response.Value.Next);
-            CheckTransactions(response.Value.Transactions, 1);
+            var response = await bruceClient.Client.GetTransactions(gbpId, bruceAccount);
+            CheckTransactionResponse(response);
         }
 
         [TestMethod]
-        public async Task ShouldFetchNextTransactions()
+        public async Task ShouldFetchQuantifiableDebitTransactions()
         {
-            var horseId = _test.CreateUniqueId("horse");
-            await _test.Client.CreateUniqueAssetType(horseId);
-            
-            var trevorId = _test.CreateUniqueId("trevor");
-            var _ = await _test.Client.CreateUniqueAsset(trevorId, horseId);
-            
             var gbpId = _test.CreateUniqueId("GBP");
-            await _test.Client.CreateQuantifiableAssetType(gbpId, 2);
+            var writeResult = await _test.Client.CreateQuantifiableAssetType(gbpId, 2);
+            Assert.IsTrue(writeResult.Success);
             
-            var account = _test.CreateUniqueId("AccountGBP");
-            await _test.Client.CreateQuantifiableAccount(account, gbpId, 1000);
+            var account = _test.CreateUniqueId("TestAccountGBP");
+            writeResult = await _test.Client.CreateQuantifiableAccount(account, gbpId, 1000);
+            Assert.IsTrue(writeResult.Success);
             
-            var __ = await _test.Client.AddBalance(account, gbpId, 50);
+            writeResult = await _test.Client.AddBalance(account, gbpId, 50);
+            Assert.IsTrue(writeResult.Success);
             
             var bruceClient = new IntegrationTestCreation();
-            var bruceAccount = bruceClient.CreateUniqueId("AccountGBP");
-            await bruceClient.Client.CreateQuantifiableAccount(bruceAccount, gbpId, 1000);
+            var bruceAccount = bruceClient.CreateUniqueId("BruceAccountGBP");
+            writeResult = await bruceClient.Client.CreateQuantifiableAccount(bruceAccount, gbpId, 1000);
+            Assert.IsTrue(writeResult.Success);
             
-            var uniqueTransfer = _test.Client.CreateOwnershipTransfer(trevorId, horseId, _test.Identity.Id, bruceClient.Identity.Id);
-            var quantity1Transfer = _test.Client.CreateQuantityTransfer(account, bruceAccount, gbpId, 5);
-            var quantity2Transfer = _test.Client.CreateQuantityTransfer(account, bruceAccount, gbpId, 10);
-            var body = new TransfersBody(quantity1Transfer, quantity2Transfer, uniqueTransfer);
+            var quantityTransfer = _test.Client.CreateQuantityTransfer(account, bruceAccount, gbpId, 10);
+            var body = new TransfersBody(quantityTransfer);
             var bodyText = body.Serialize();
             var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
                 new[]
@@ -102,24 +90,218 @@ namespace IntegrationTests
                     _test.Client.GenerateAuthorisation(bodyText),
                     bruceClient.Client.GenerateAuthorisation(bodyText)
                 });
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+            
+            var response = await _test.Client.GetTransactions(gbpId, account);
+            CheckTransactionResponse(response);
+        }
 
-            var ____ = await _test.Client.Write(transferRequest);
+        [TestMethod]
+        public async Task ShouldFetchUniqueCreditTransactions()
+        {
+            var horseId = _test.CreateUniqueId("horse");
+            var writeResult = await _test.Client.CreateUniqueAssetType(horseId);
+            Assert.IsTrue(writeResult.Success);
             
-            // Currently only supports credits
-            // var response = await _test.Client.GetTransactions(horseId, trevorId, 1, null);
+            var trevorId = _test.CreateUniqueId("trevor");
+            writeResult = await _test.Client.CreateUniqueAsset(trevorId, horseId);
+            Assert.IsTrue(writeResult.Success);
+            
+            var bruceClient = new IntegrationTestCreation();
+            var uniqueTransfer = _test.Client.CreateOwnershipTransfer(trevorId, horseId, _test.Identity.Id, bruceClient.Identity.Id);
+            var body = new TransfersBody(uniqueTransfer);
+            var bodyText = body.Serialize();
+            var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
+                new[]
+                {
+                    _test.Client.GenerateAuthorisation(bodyText),
+                    bruceClient.Client.GenerateAuthorisation(bodyText)
+                });
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+
+            var response = await bruceClient.Client.GetTransactions(horseId, trevorId);
+            CheckTransactionResponse(response);
+        }
+
+        [TestMethod]
+        public async Task ShouldFetchUniqueDebitTransactions()
+        {
+            var horseId = _test.CreateUniqueId("horse");
+            var writeResult = await _test.Client.CreateUniqueAssetType(horseId);
+            Assert.IsTrue(writeResult.Success);
+            
+            var trevorId = _test.CreateUniqueId("trevor");
+            writeResult = await _test.Client.CreateUniqueAsset(trevorId, horseId);
+            Assert.IsTrue(writeResult.Success);
+            
+            var bruceClient = new IntegrationTestCreation();
+            var uniqueTransfer = _test.Client.CreateOwnershipTransfer(trevorId, horseId, _test.Identity.Id, bruceClient.Identity.Id);
+            var body = new TransfersBody(uniqueTransfer);
+            var bodyText = body.Serialize();
+            var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
+                new[]
+                {
+                    _test.Client.GenerateAuthorisation(bodyText),
+                    bruceClient.Client.GenerateAuthorisation(bodyText)
+                });
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+            
+            var response = await _test.Client.GetTransactions(horseId, trevorId);
+            CheckTransactionResponse(response);
+        }
+
+        [TestMethod]
+        public async Task ShouldFetchNextQuantifiableTransactions()
+        {
+            var gbpId = _test.CreateUniqueId("GBP");
+            var writeResult = await _test.Client.CreateQuantifiableAssetType(gbpId, 2);
+            Assert.IsTrue(writeResult.Success);
+            
+            var account = _test.CreateUniqueId("AccountGBP");
+            writeResult = await _test.Client.CreateQuantifiableAccount(account, gbpId, 1000);
+            Assert.IsTrue(writeResult.Success);
+            
+            writeResult = await _test.Client.AddBalance(account, gbpId, 50);
+            Assert.IsTrue(writeResult.Success);
+            
+            var bruceClient = new IntegrationTestCreation();
+            var bruceAccount = bruceClient.CreateUniqueId("AccountGBP");
+            writeResult = await bruceClient.Client.CreateQuantifiableAccount(bruceAccount, gbpId, 1000);
+            Assert.IsTrue(writeResult.Success);
+            
+            var quantity1Transfer = _test.Client.CreateQuantityTransfer(account, bruceAccount, gbpId, 5);
+            var quantity2Transfer = _test.Client.CreateQuantityTransfer(account, bruceAccount, gbpId, 10);
+            var body = new TransfersBody(quantity1Transfer, quantity2Transfer);
+            var bodyText = body.Serialize();
+            var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
+                new[]
+                {
+                    _test.Client.GenerateAuthorisation(bodyText),
+                    bruceClient.Client.GenerateAuthorisation(bodyText)
+                });
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+            
             var response = await _test.Client.GetTransactions(gbpId, account, 1);
-            
-            Assert.IsNotNull(response);
-            Assert.IsTrue(response.Success);
-            Assert.IsNotNull(response.Value.Next);
-            CheckTransactions(response.Value.Transactions, 1);
+            CheckTransactionResponse(response, 1, true);
 
             response = await _test.Client.GetTransactions(gbpId, account, 1, response.Value.Next);
+            CheckTransactionResponse(response);            
+        }
+
+        [TestMethod]
+        public async Task ShouldFetchNextUniqueTransactions()
+        {
+            var horseId = _test.CreateUniqueId("horse");
+            var writeResult = await _test.Client.CreateUniqueAssetType(horseId);
+            Assert.IsTrue(writeResult.Success);
             
+            var trevorId = _test.CreateUniqueId("trevor");
+            writeResult = await _test.Client.CreateUniqueAsset(trevorId, horseId);
+            Assert.IsTrue(writeResult.Success);
+
+            var bruceClient = new IntegrationTestCreation();
+            
+            var uniqueTransfer1 = _test.Client.CreateOwnershipTransfer(trevorId, horseId, _test.Identity.Id, bruceClient.Identity.Id);
+            var body = new TransfersBody(uniqueTransfer1);
+            var bodyText = body.Serialize();
+            var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
+                new[]
+                {
+                    _test.Client.GenerateAuthorisation(bodyText)
+                });
+
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+
+            var uniqueTransfer2 = bruceClient.Client.CreateOwnershipTransfer(trevorId, horseId, bruceClient.Identity.Id, _test.Identity.Id);
+            body = new TransfersBody(uniqueTransfer2);
+            bodyText = body.Serialize();
+            transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
+                new[]
+                {
+                    bruceClient.Client.GenerateAuthorisation(bodyText)
+                });
+
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+
+            var response = await _test.Client.GetTransactions(horseId, trevorId, 1);
+            CheckTransactionResponse(response, 1, true);
+            response = await _test.Client.GetTransactions(horseId, trevorId, 1, response.Value.Next);
+            CheckTransactionResponse(response);
+        }
+
+        [TestMethod]
+        public async Task ShouldDelegatesSeeHistory()
+        {
+            // Create asset type
+            var gbpId = _test.CreateUniqueId("GBP");
+            await _test.Client.CreateQuantifiableAssetType(gbpId, 2);
+            
+            var delegate1 = _test.IdentityBuilder.Create();
+            var writeResult = await _test.Client.CreateIdentity(delegate1);
+            Assert.IsTrue(writeResult.Success);
+
+            writeResult = await _test.Client.AddDelegate(delegate1);
+            Assert.IsTrue(writeResult.Success);
+
+            using var delegateClient1 = await ClientBuilder.CreateWithNewIdentity(TestEnvironment.Environment, delegate1);
+            delegateClient1.UseDelegator(_test.Identity.Id);
+
+            var delegate2 = _test.IdentityBuilder.Create();
+            writeResult = await _test.Client.CreateIdentity(delegate2);
+            Assert.IsTrue(writeResult.Success);
+            writeResult = await _test.Client.AddDelegate(delegate2);
+            Assert.IsTrue(writeResult.Success);
+            using var delegateClient2 = await ClientBuilder.CreateWithNewIdentity(TestEnvironment.Environment, delegate2);
+            delegateClient2.UseDelegator(_test.Identity.Id);
+
+            var accountA = _test.CreateUniqueId("AccountA");
+            writeResult = await delegateClient1.CreateQuantifiableAccount(accountA, gbpId);
+            Assert.IsTrue(writeResult.Success);
+            writeResult = await delegateClient1.AddBalance(accountA, gbpId, 10);
+            Assert.IsTrue(writeResult.Success);
+
+            var accountB = _test.CreateUniqueId("AccountB");
+            writeResult = await delegateClient1.CreateQuantifiableAccount(accountB, gbpId);
+            Assert.IsTrue(writeResult.Success);
+
+            var quantity1Transfer = delegateClient1.CreateQuantityTransfer(accountA, accountB, gbpId, 3);
+            var body = new TransfersBody(quantity1Transfer);
+            var bodyText = body.Serialize();
+            var transferRequest = new PlatformWriteRequest(body.RequestId, bodyText,
+                new[]
+                {
+                    delegateClient1.GenerateAuthorisation(bodyText)
+                });
+            writeResult = await _test.Client.Write(transferRequest);
+            Assert.IsTrue(writeResult.Success);
+
+            var getQuantifiableAssetResponse = await delegateClient2.GetQuantifiableAsset(accountA, gbpId);
+            Assert.AreEqual("7", getQuantifiableAssetResponse.Value.Quantity);
+
+            getQuantifiableAssetResponse = await delegateClient2.GetQuantifiableAsset(accountB, gbpId);
+            Assert.AreEqual("3", getQuantifiableAssetResponse.Value.Quantity);
+
+            var response = await delegateClient2.GetTransactions(gbpId, accountA);
+            CheckTransactionResponse(response);
+            response = await delegateClient2.GetTransactions(gbpId, accountB);
+            CheckTransactionResponse(response);
+        }
+
+        private static void CheckTransactionResponse(ResponseResult<TransactionsResult> response, int expected = 1, bool expectedNext = false)
+        {
             Assert.IsNotNull(response);
             Assert.IsTrue(response.Success);
-            Assert.IsNull(response.Value.Next);
-            CheckTransactions(response.Value.Transactions, 1);
+            if (expectedNext)
+                Assert.IsNotNull(response.Value.Next);
+            else
+                Assert.IsNull(response.Value.Next);
+            CheckTransactions(response.Value.Transactions, expected);
         }
 
         private static void CheckTransactions(IReadOnlyCollection<Transaction> transactions, int expected)
@@ -128,16 +310,29 @@ namespace IntegrationTests
             Assert.AreEqual(expected, transactions.Count);
             foreach (var transaction in transactions)
             {
+
                 Assert.IsNotNull(transaction.Proof);
                 Assert.IsNotNull(transaction.RequestId);
                 Assert.IsNotNull(transaction.TransactionTimestamp);
-                Assert.IsNotNull(transaction.Quantity);
-                Assert.IsNotNull(transaction.Sender);
-                Assert.IsNotNull(transaction.Sender.AssetId);
-                Assert.IsNotNull(transaction.Sender.AssetTypeId);
-                Assert.IsNotNull(transaction.Recipient);
-                Assert.IsNotNull(transaction.Recipient.AssetId);
-                Assert.IsNotNull(transaction.Recipient.AssetTypeId);
+                switch (transaction)
+                {
+                    case QuantifiableTransaction quantifiable:
+                        Assert.IsNotNull(quantifiable.Quantity);
+                        Assert.IsNotNull(quantifiable.Sender);
+                        Assert.IsNotNull(quantifiable.Sender.AssetId);
+                        Assert.IsNotNull(quantifiable.Sender.AssetTypeId);
+                        Assert.IsNotNull(quantifiable.Recipient);
+                        Assert.IsNotNull(quantifiable.Recipient.AssetId);
+                        Assert.IsNotNull(quantifiable.Recipient.AssetTypeId);
+                        break;
+                    case UniqueTransaction unique:
+                        Assert.IsNotNull(unique.Asset);
+                        Assert.IsNotNull(unique.Asset.AssetId);
+                        Assert.IsNotNull(unique.Asset.AssetTypeId);
+                        Assert.IsNotNull(unique.FromOwner);
+                        Assert.IsNotNull(unique.ToOwner);
+                        break;
+                }
             }
         }
     }