+++ Iov42sdk/Models/Headers/AuthenticationHeader.cs
@@ -7,15 +7,17 @@
         {
         }
 
-        public AuthenticationHeader(string protocolId, string identityId, string signature)
+        public AuthenticationHeader(string protocolId, string identityId, string signature, string delegateIdentityId = null)
         {
             IdentityId = identityId;
             ProtocolId = protocolId;
             Signature = signature;
+            DelegateIdentityId = delegateIdentityId;
         }
 
         public string IdentityId { get; set; }
         public string ProtocolId { get; set; }
         public string Signature { get; set; }
+        public string DelegateIdentityId { get; set; }
     }
 }