+++ Iov42sdk/Models/Transactions/CombinedTransactionsResult.cs
@@ -1,0 +1,8 @@
+﻿namespace Iov42sdk.Models.Transactions
+{
+    public class CombinedTransactionsResult
+    {
+        public string Next { get; set; }
+        public CombinedTransaction[] Transactions { get; set; }
+    }
+}